extract to mapper dir
