+++ src/modules/0.0.22/aws_cloudwatch/mappers/index.ts
@@ -1,0 +1,1 @@
+export * from './log_group';
